refactor into module functions to help with maintainability
